requirements.txt
streamlit
langchain
langgraph
openai
python-dotenv
rdkit
aiohttp
